+++ background_subtraction/GMM/README.md
@@ -1,0 +1,6 @@
+# Mixture of Gaussians
+
+1. **MOG.py:** Use for extracting foreground and apply morphological filters to each frame.
+- We can visualize the results (display a video) using script visualize.py.
+
+2. **tracking.py:** Use for performing tracking and visualize it.
