--- background_subtraction/GMM/Tracking.py
@@ -1,70 +1,0 @@
-import numpy as np
-from scipy.spatial import distance as dist
-from collections import OrderedDict
-
-#  Reference: https://www.pyimagesearch.com/2018/07/23/simple-object-tracking-with-opencv/
-
-class CentroidTracker():
-    def __init__(self, maxDisappeared = 15):
-        self.nextObjectID = 0  # counter used to assign unique IDs to each object
-        self.objects = OrderedDict()  # Object ID - Centroid(x,y)
-        self.disappeared = OrderedDict()  # Object ID - #frames where object is lost
-        self.maxDisappeared = maxDisappeared
-
-    def register(self, centroid):
-        self.objects[self.nextObjectID] = centroid
-        self.disappeared[self.nextObjectID] = 0
-        self.nextObjectID += 1
-
-    def deregister(self, objectID):
-        del self.objects[objectID]
-        del self.disappeared[objectID]
-
-    def update(self, rects):  # rects is a tuple (startX, startY, endX, endY)
-        if len(rects) == 0:
-            for objectID in self.disappeared.keys():
-                self.disappeared[objectID] += 1
-                if self.disappeared[objectID] > self.maxDisappeared:
-                    self.deregister(objectID)
-            return self.objects
-
-        inputCentroids = np.zeros((len(rects), 2), dtype="int")
-        # TODO: Aqui posar els meus centroids
-        for (i, (startX, startY, endX, endY)) in enumerate(rects):
-            cX = int((startX + endX) / 2.0)
-            cY = int((startY + endY) / 2.0)
-            inputCentroids[i] = (cX, cY)
-
-        if len(self.objects) == 0:
-            for i in range(0, len(inputCentroids)):
-                self.register(inputCentroids[i])
-        # Match the input centroids to existing object centroids
-        else:
-            objectIDs = list(self.objects.keys())
-            objectCentroids = list(self.objects.values())
-            D = dist.cdist(np.array(objectCentroids), inputCentroids)
-            rows = D.min(axis=1).argsort()
-            cols = D.argmin(axis=1)[rows]
-            usedRows = set()
-            usedCols = set()
-            for (row, col) in zip(rows, cols):
-                if row in usedRows or col in usedCols:
-                    continue
-                objectID = objectIDs[row]
-                self.objects[objectID] = inputCentroids[col]
-                self.disappeared[objectID] = 0
-                usedRows.add(row)
-                usedCols.add(col)
-            unusedRows = set(range(0, D.shape[0])).difference(usedRows)
-            unusedCols = set(range(0, D.shape[1])).difference(usedCols)
-            if D.shape[0] >= D.shape[1]:
-                # loop over the unused row indexes
-                for row in unusedRows:
-                    objectID = objectIDs[row]
-                    self.disappeared[objectID] += 1
-                    if self.disappeared[objectID] > self.maxDisappeared:
-                        self.deregister(objectID)
-            else:
-                for col in unusedCols:
-                    self.register(inputCentroids[col])
-        return self.objects
