+++ background_subtraction/GMM/testXML.py
@@ -1,0 +1,16 @@
+import xml.etree.ElementTree as ET
+
+# create the file structure
+data = ET.Element('data')
+items = ET.SubElement(data, 'items')
+item1 = ET.SubElement(items, 'item')
+item2 = ET.SubElement(items, 'item')
+item1.set('name','item1')
+item2.set('name','item2')
+item1.text = 'item1abc'
+item2.text = 'item2abc'
+
+# create a new XML file with the results
+mydata = ET.tostring(data)
+myfile = open("items2.xml", "w")
+myfile.write(mydata)
